move inspectr app into plottr
