+++ plottr/apps/inspectr.py
@@ -14,6 +14,8 @@ data. it relies on the public qcodes API to get its information.
 
 import os
 import time
+import sys
+import argparse
 
 from pyqtgraph.Qt import QtGui, QtCore
 
@@ -478,3 +480,22 @@ class QCodesDBInspector(QtGui.QMainWindow):
 def inspectr(dbPath: str = None):
     win = QCodesDBInspector(dbPath=dbPath)
     return win
+
+
+def main(dbPath):
+    app = QtGui.QApplication([])
+    plottrlog.enableStreamHandler(True)
+
+    win = inspectr(dbPath=dbPath)
+    win.show()
+
+    if (sys.flags.interactive != 1) or not hasattr(QtCore, 'PYQT_VERSION'):
+        QtGui.QApplication.instance().exec_()
+
+
+def script():
+    parser = argparse.ArgumentParser(description='inspectr -- sifting through qcodes data.')
+    parser.add_argument('--dbpath', help='path to qcodes .db file',
+                        default=None)
+    args = parser.parse_args()
+    main(args.dbpath)
